requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pydantic
httpx

